[Spring 지하철 노선도 - 3단계] 알파(배철환) 미션 제출합니다. (#331)
* docs : 기능요구사항 추가

* docs : 도메인 추가 및 기능요구사항 추가

* feat : 구간을 추가하고 삭제하는 기능을 담당하는 Stations 추가

* feat : 한 노선의 상행 종점과 하행 종점을 찾는 기능 추가

* feat : SectionDao의 기능 구현

* feat : 구간을 추가하고 삭제하는 SectionService의 기능 추가

* feat : LineService에 노선 추가, 삭제, 수정, 조회, 목록 기능 추가

* feat : StationService에 역 추가, 삭제, 조회 기능 추가

* feat : SectionController의 기능 추가

* feat : LineController의 기능 추가

* refactor : 구간 삭제와 구간 삽입의 비즈니스 로직을 담당하는 Sections 추가

* refactor : 삭제된 MetroManager를 의존하던 클래스와 테스트 케이스 수정

* refactor : 역을 삭제하는 경우 내부 로직 수정

* refactor : 클래스명 변경

* refactor : SectionDao의 메소드명 변경

* refactor : DB에서 특정 데이터가 있는지 조회할 때 count가 아닌 exist로 수정

* refactor : 테스트 코드 내의 사용하지 않는 지역변수 제거

* refactor : 커스텀 예외가 아닌 예외에 대한 Response를 추가

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -4,16 +4,26 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.SubwayException;
+import wooteco.subway.dao.LineDao;
+import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
+import wooteco.subway.domain.Line;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
 
 @Service
+@Transactional
 public class StationService {
     private final StationDao stationDao;
+    private final SectionDao sectionDao;
+    private final LineDao lineDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(LineDao lineDao, SectionDao sectionDao, StationDao stationDao) {
+        this.lineDao = lineDao;
+        this.sectionDao = sectionDao;
         this.stationDao = stationDao;
     }
 
@@ -32,12 +42,15 @@ public class StationService {
 
     private void validateId(Long id) {
         if (!stationDao.isValidId(id)) {
-            throw new EmptyResultDataAccessException(id.intValue());
+            throw new SubwayException("[ERROR] 존재하지 않는 역을 삭제할 수 없습니다.");
         }
     }
 
     public void deleteById(Long id) {
         validateId(id);
+        if (sectionDao.isAnySectionWithStationId(id)) {
+            throw new SubwayException("[ERROR] 구간을 구성하는 역은 삭제할 수 없습니다.");
+        }
         stationDao.deleteById(id);
     }
 }
